+++ Client/Client.cs
@@ -61,6 +61,7 @@ public partial class Client : IDisposable, IInputManagement
     private readonly ConsoleCommands m_consoleCommands = new();
     private readonly Profiler m_profiler = new();
     private readonly Ticker m_ticker = new(Constants.TicksPerSecond);
+    private readonly SaveGameScreenshotGenerator m_screenshotGenerator;
     private bool m_disposed;
     private bool m_takeScreenshot;
     private bool m_loadComplete;
@@ -99,11 +100,12 @@ public partial class Client : IDisposable, IInputManagement
         }
 
         m_window = new Window(AppInfo.ApplicationName, config, archiveCollection, m_fpsTracker, this, GlVersion.Major, GlVersion.Minor, GlVersion.Flags, CheckOpenGLSupport);
+        m_screenshotGenerator = new(m_window.Renderer);
         m_soundManager.SoundCreated += m_window.JoystickAdapter.RumbleForSoundCreated;
         SetIcon(m_window);
 
         m_layerManager = new GameLayerManager(config, m_window, console, m_consoleCommands, archiveCollection,
-            m_soundManager, m_saveGameManager, m_profiler);
+            m_soundManager, m_saveGameManager, m_profiler, m_screenshotGenerator);
 
         m_layerManager.GameLayerAdded += GameLayerManager_GameLayerAdded;
         m_saveGameManager.GameSaved += SaveGameManager_GameSaved;
@@ -268,6 +270,7 @@ public partial class Client : IDisposable, IInputManagement
 
         m_filesLoaded = false;
         m_window.Renderer.UploadColorMap();
+        m_saveGameManager.LoadCurrentSaveFiles();
     }
 
     private void CheckMapLoad()
@@ -335,7 +338,7 @@ public partial class Client : IDisposable, IInputManagement
         // Note: StaticDataApplier happens through this start and needs to happen before UpdateToNewWorld
         worldLayer.World.Start(m_loadMapResult.WorldModel);
 
-        WriteAutoSave(m_loadMapResult);
+        _ = WriteAutoSave(m_loadMapResult);
 
         m_window.Renderer.UpdateToNewWorld(worldLayer.World);
         m_layerManager.LockInput = false;
@@ -354,7 +357,7 @@ public partial class Client : IDisposable, IInputManagement
         m_onLoadMapComplete = null;
     }
 
-    private void WriteAutoSave(LoadMapResult result)
+    private async Task WriteAutoSave(LoadMapResult result)
     {
         if (result.WorldLayer == null || result.Players.Count == 0 || !m_config.Game.AutoSave)
             return;
@@ -363,7 +366,7 @@ public partial class Client : IDisposable, IInputManagement
         var mapInfoDef = worldLayer.CurrentMap;
 
         string title = $"Auto: {mapInfoDef.GetMapNameWithPrefix(worldLayer.World.ArchiveCollection.Language)}";
-        var saveGameEvent = m_saveGameManager.WriteNewSaveGame(worldLayer.World, title, autoSave: true);
+        var saveGameEvent = await m_saveGameManager.WriteNewSaveGameAsync(worldLayer.World, title, m_screenshotGenerator, autoSave: true);
         if (saveGameEvent.Success)
             m_console.AddMessage($"Saved {saveGameEvent.FileName}");