+++ Core/Graphics/ImageExtensions.cs
@@ -1,5 +1,8 @@
-﻿using SixLabors.ImageSharp;
+﻿using Helion.Geometry;
+using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.Drawing;
 using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.Processing;
 using System.IO;
 
 namespace Helion.Graphics;
@@ -8,27 +11,56 @@ public static class ImageExtensions
 {
     public static bool SavePng(this Image image, string path)
     {
+        using FileStream fs = new(path, FileMode.CreateNew);
+        return SavePng(image, fs, null);
+    }
+
+    public static bool SavePng(this Image image, Stream stream, Dimension? resize)
+    {
         try
         {
             var pixels = image.Pixels;
             byte[] data = new byte[pixels.Length * 4]; // rgba -> [r, g, b]
-            for (int i = 0; i < pixels.Length; i++)
-            {
-                uint pixel = pixels[i];
-                byte r = (byte)((pixel & 0x00FF0000) >> 16);
-                byte g = (byte)((pixel & 0x0000FF00) >> 8);
-                byte b = (byte)(pixel & 0x000000FF);
 
-                int offset = i * 4;
-                data[offset] = r;
-                data[offset + 1] = g;
-                data[offset + 2] = b;
-                data[offset + 3] = 255;
+            if (image.ImageType == ImageType.Rgba)
+            {
+                for (int i = 0; i < pixels.Length; i++)
+                {
+                    uint pixel = pixels[i];
+                    int offset = i * 4;
+                    data[offset] = (byte)((pixel) & 0xFF);
+                    data[offset + 1] = (byte)((pixel >> 8) & 0xFF);
+                    data[offset + 2] = (byte)((pixel >> 16) & 0xFF);
+                    data[offset + 3] = 255;
+                }
+            }
+            else
+            {
+                for (int i = 0; i < pixels.Length; i++)
+                {
+                    uint pixel = pixels[i];
+                    int offset = i * 4;
+                    data[offset] = (byte)((pixel >> 16) & 0xFF);
+                    data[offset + 1] = (byte)((pixel >> 8) & 0xFF);
+                    data[offset + 2] = (byte)((pixel) & 0xFF);
+                    data[offset + 3] = 255;
+                }
             }
 
             using var pixelImage = SixLabors.ImageSharp.Image.LoadPixelData<Rgba32>(data, image.Width, image.Height);
-            using FileStream fs = new(path, FileMode.CreateNew);
-            pixelImage.SaveAsPng(fs, null);
+            if (resize.HasValue)
+            {
+                var resizeOptions = new ResizeOptions
+                {
+                    Size = new Size(resize.Value.Width, resize.Value.Height),
+                    Mode = ResizeMode.Crop,
+                    PadColor = SixLabors.ImageSharp.Color.Black
+                };
+
+                pixelImage.Mutate(x => x.Resize(resizeOptions));
+            }
+
+            pixelImage.SaveAsPng(stream, null);
             return true;
         }
         catch