+++ Core/Layer/GameLayerManager.cs
@@ -36,6 +36,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Threading.Tasks;
 using static Helion.Util.Assertion.Assert;
 
 namespace Helion.Layer;
@@ -80,6 +81,7 @@ public class GameLayerManager : IGameLayerManager
     private readonly HudRenderContext m_hudContext = new(default);
     private readonly OptionsLayer m_optionsLayer;
     private readonly ConsoleLayer m_consoleLayer;
+    private readonly IScreenshotGenerator m_screenshotGenerator;
     private Renderer m_renderer;
     private IRenderableSurfaceContext m_ctx;
     private IHudRenderContext m_hudRenderCtx;
@@ -93,10 +95,11 @@ public class GameLayerManager : IGameLayerManager
 
     public GameLayerManager(IConfig config, IWindow window, HelionConsole console, ConsoleCommands consoleCommands,
         ArchiveCollection archiveCollection, SoundManager soundManager, SaveGameManager saveGameManager,
-        Profiler profiler)
+        Profiler profiler, IScreenshotGenerator screenshotGenerator)
     {
         m_config = config;
         m_window = window;
+        m_screenshotGenerator = screenshotGenerator;
         m_console = console;
         m_consoleCommands = consoleCommands;
         m_archiveCollection = archiveCollection;
@@ -248,7 +251,7 @@ public class GameLayerManager : IGameLayerManager
         if (m_quickSaveTicks >= seconds * (int)Constants.TicksPerSecond)
         {
             m_quickSaveTicks = 0;
-            WriteQuickSave();
+            _ = WriteQuickSave();
         }
     }
 
@@ -489,7 +492,7 @@ public class GameLayerManager : IGameLayerManager
         }
 
         if (ConsumeCommandPressed(Constants.Input.QuickSave, input))
-            QuickSave();
+            _ = QuickSave();
 
         if (ConsumeCommandPressed(Constants.Input.Load, input))
             GoToSaveOrLoadMenu(false);
@@ -587,7 +590,7 @@ public class GameLayerManager : IGameLayerManager
 
         if (MenuLayer == null)
         {
-            MenuLayer menuLayer = new(this, m_config, m_console, m_archiveCollection, m_soundManager, m_saveGameManager, m_optionsLayer);
+            MenuLayer menuLayer = new(this, m_config, m_console, m_archiveCollection, m_soundManager, m_saveGameManager, m_optionsLayer, m_screenshotGenerator);
             menuLayer.Animation.AnimateIn();
             Add(menuLayer);
         }
@@ -622,7 +625,7 @@ public class GameLayerManager : IGameLayerManager
         Add(new EndoomLayer(closeAction, m_archiveCollection, m_window.Dimension.Height));
     }
 
-    public void QuickSave()
+    public async Task QuickSave()
     {
         if (!CanSave)
             return;
@@ -630,7 +633,7 @@ public class GameLayerManager : IGameLayerManager
         // if we're using rotating quicksaves, then we aren't concerned with saving to a particular slot
         if (m_config.Game.RotatingQuickSaves > 0)
         {
-            WriteQuickSave();
+            await WriteQuickSave();
             return;
         }
 
@@ -643,7 +646,7 @@ public class GameLayerManager : IGameLayerManager
         if (m_config.Game.QuickSaveConfirm)
         {
             MessageMenu confirm = new(m_config, m_console, m_soundManager, m_archiveCollection,
-                new[] { "Are you sure you want to overwrite:", LastSave?.SaveGame.Model != null ? LastSave.Value.SaveGame.Model.Text : "Save", "Press Y to confirm." },
+                ["Are you sure you want to overwrite:", LastSave?.SaveGame.Model != null ? LastSave.Value.SaveGame.Model.Text : "Save", "Press Y to confirm."],
                 isYesNoConfirm: true, clearMenus: true);
             confirm.Cleared += Confirm_Cleared;
 
@@ -652,7 +655,7 @@ public class GameLayerManager : IGameLayerManager
             return;
         }
 
-        WriteQuickSave();
+        await WriteQuickSave();
     }
 
     private void Confirm_Cleared(object? sender, bool e)
@@ -660,10 +663,10 @@ public class GameLayerManager : IGameLayerManager
         if (!e || !LastSave.HasValue)
             return;
 
-        WriteQuickSave();
+        _ = WriteQuickSave();
     }
 
-    private void WriteQuickSave()
+    private async Task WriteQuickSave()
     {
         bool isRotating = m_config.Game.RotatingQuickSaves > 0;
         if (WorldLayer == null || (!isRotating && LastSave == null) || !CanSave)
@@ -673,7 +676,7 @@ public class GameLayerManager : IGameLayerManager
         if (isRotating)
         {
             string name = $"Quick: {world.MapInfo.GetMapNameWithPrefix(world.ArchiveCollection.Language)}";
-            var saveEvent = m_saveGameManager.WriteSaveGame(world, name, null, quickSave: true);
+            var saveEvent = await m_saveGameManager.WriteSaveGameAsync(world, name, m_screenshotGenerator, null, quickSave: true);
             HandleSaveEvent(saveEvent, world);
         }
         else
@@ -684,15 +687,17 @@ public class GameLayerManager : IGameLayerManager
             string name = isCustomizedName
                 ? existingSave.Model?.Text ?? "Unnamed"
                 : world.MapInfo.GetMapNameWithPrefix(world.ArchiveCollection.Language);
-            var saveEvent = m_saveGameManager.WriteSaveGame(world, name, existingSave);
+            var saveEvent = await m_saveGameManager.WriteSaveGameAsync(world, name, m_screenshotGenerator, existingSave);
             HandleSaveEvent(saveEvent, world, SaveMenu.SaveMessage);
         }
     }
 
-    private void HandleSaveEvent(SaveGameEvent saveEvent, SinglePlayerWorld world, string? successMessage = null)
+    private static void HandleSaveEvent(SaveGameEvent saveEvent, SinglePlayerWorld world, string? successMessage = null)
     {
         if (saveEvent.Success)
+        {
             world.DisplayMessage(world.Player, null, successMessage ?? $"Saved {saveEvent.FileName}");
+        }
         else
         {
             world.DisplayMessage(world.Player, null, $"Failed to save {saveEvent.FileName}");