+++ Core/Layer/Menus/MenuLayer.cs
@@ -7,6 +7,7 @@ using Helion.Menus;
 using Helion.Menus.Impl;
 using Helion.Render.Common.Renderers;
 using Helion.Resources.Archives.Collection;
+using Helion.Util;
 using Helion.Util.Configs;
 using Helion.Util.Consoles;
 using Helion.Util.Timing;
@@ -26,11 +27,12 @@ public partial class MenuLayer : IGameLayer, IAnimationLayer
     private readonly Stack<Menu> m_menus = new();
     private readonly Stopwatch m_stopwatch = new();
     private readonly OptionsLayer m_optionsLayer;
+    private readonly IScreenshotGenerator m_screenshotGenerator;
     private readonly Action<IHudRenderContext> m_renderVirtualHudAction;
     private bool m_disposed;
 
     public MenuLayer(GameLayerManager manager, IConfig config, HelionConsole console,
-        ArchiveCollection archiveCollection, SoundManager soundManager, SaveGameManager saveGameManager, OptionsLayer optionsLayer)
+        ArchiveCollection archiveCollection, SoundManager soundManager, SaveGameManager saveGameManager, OptionsLayer optionsLayer, IScreenshotGenerator screenshotGenerator)
     {
         Manager = manager;
         m_config = config;
@@ -39,12 +41,13 @@ public partial class MenuLayer : IGameLayer, IAnimationLayer
         m_soundManager = soundManager;
         m_saveGameManager = saveGameManager;
         m_optionsLayer = optionsLayer;
+        m_screenshotGenerator = screenshotGenerator;
         m_renderVirtualHudAction = new(RenderVirtualHud);
         m_stopwatch.Start();
 
         Animation = new(TimeSpan.FromMilliseconds(200), this);
 
-        MainMenu mainMenu = new(this, config, console, soundManager, archiveCollection, saveGameManager, optionsLayer);
+        MainMenu mainMenu = new(this, config, console, soundManager, archiveCollection, saveGameManager, optionsLayer, m_screenshotGenerator);
         m_menus.Push(mainMenu);
     }
 
@@ -61,7 +64,7 @@ public partial class MenuLayer : IGameLayer, IAnimationLayer
 
         bool hasWorld = Manager.WorldLayer != null;
         SaveMenu saveMenu = new(this, m_config, m_console, m_soundManager, m_archiveCollection,
-            m_saveGameManager, hasWorld, isSave, clearOnExit);
+            m_saveGameManager, m_screenshotGenerator, hasWorld, isSave, clearOnExit);
 
         m_menus.Push(saveMenu);
     }
@@ -81,7 +84,7 @@ public partial class MenuLayer : IGameLayer, IAnimationLayer
     public void SetToMainMenu()
     {
         m_menus.Clear();
-        m_menus.Push(new MainMenu(this, m_config, m_console, m_soundManager, m_archiveCollection, m_saveGameManager, m_optionsLayer));
+        m_menus.Push(new MainMenu(this, m_config, m_console, m_soundManager, m_archiveCollection, m_saveGameManager, m_optionsLayer, m_screenshotGenerator));
     }
 
     public void Close()