+++ Core/Util/IScreenshotGenerator.cs
@@ -1,0 +1,8 @@
+﻿using Helion.Graphics;
+namespace Helion.Util;
+
+public interface IScreenshotGenerator
+{
+    Image? GetImage();
+    byte[]? GeneratePngImage(Image image);
+}