+++ Core/World/Save/SaveGameManager.cs
@@ -2,8 +2,11 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Threading.Tasks;
+using Helion.Graphics;
 using Helion.Models;
 using Helion.Resources.Archives.Collection;
+using Helion.Util;
 using Helion.Util.Configs;
 using Helion.Util.Extensions;
 using Helion.World.Util;
@@ -11,39 +14,25 @@ using NLog;
 
 namespace Helion.World.Save;
 
-public readonly struct SaveGameEvent
+public readonly struct SaveGameEvent(SaveGame saveGame, WorldModel worldModel, string filename, bool success, Exception? ex = null)
 {
-    public readonly SaveGame SaveGame;
-    public readonly WorldModel WorldModel;
-    public readonly string FileName;
-    public readonly bool Success;
-    public readonly Exception? Exception;
-
-    public SaveGameEvent(SaveGame saveGame, WorldModel worldModel, string filename, bool success, Exception? ex = null)
-    {
-        SaveGame = saveGame;
-        FileName = filename;
-        WorldModel = worldModel;
-        Success = success;
-        Exception = ex;
-    }
+    public readonly SaveGame SaveGame = saveGame;
+    public readonly WorldModel WorldModel = worldModel;
+    public readonly string FileName = filename;
+    public readonly bool Success = success;
+    public readonly Exception? Exception = ex;
 }
 
-public class SaveGameManager
+public class SaveGameManager(IConfig config, ArchiveCollection archiveCollection, string? saveDirCommandLineArg)
 {
     private static readonly Logger Log = LogManager.GetCurrentClassLogger();
-    private readonly IConfig m_config;
-    private readonly ArchiveCollection m_archiveCollection;
-    private readonly string? m_saveDirCommandLineArg;
+    private readonly IConfig m_config = config;
+    private readonly ArchiveCollection m_archiveCollection = archiveCollection;
+    private readonly string? m_saveDirCommandLineArg = saveDirCommandLineArg;
+    private readonly List<SaveGame> m_currentSaves = [];
+    private bool m_currentSavesLoaded;
 
     public event EventHandler<SaveGameEvent>? GameSaved;
-
-    public SaveGameManager(IConfig config, ArchiveCollection archiveCollection, string? saveDirCommandLineArg)
-    {
-        m_config = config;
-        m_archiveCollection = archiveCollection;
-        m_saveDirCommandLineArg = saveDirCommandLineArg;
-    }
 
     private string GetSaveDir()
     {
@@ -73,46 +62,90 @@ public class SaveGameManager
         }
     }
 
+    public void LoadCurrentSaveFiles()
+    {
+        if (m_currentSavesLoaded)
+            return;
+
+        var saveGames = GetMatchingSaveGames(ReadSaveGameFiles());
+        m_currentSaves.AddRange(saveGames);
+        m_currentSavesLoaded = true;
+    }
+
     public bool SaveFileExists(string filename)
     {
         string filePath = Path.Combine(GetSaveDir(), filename);
         return File.Exists(filePath);
     }
 
-    public SaveGame ReadSaveGame(string filename) => new SaveGame(GetSaveDir(), filename);
+    public SaveGame ReadSaveGame(string filename) => new(GetSaveDir(), filename);
 
-    public SaveGameEvent WriteNewSaveGame(IWorld world, string title, bool autoSave = false, bool quickSave = false) =>
-        WriteSaveGame(world, title, null, autoSave, quickSave);
+    public Task<SaveGameEvent> WriteNewSaveGameAsync(IWorld world, string title, IScreenshotGenerator screenshotGenerator, bool autoSave = false, bool quickSave = false) =>
+        WriteSaveGameAsync(world, title, screenshotGenerator, null, autoSave, quickSave);
 
-    public SaveGameEvent WriteSaveGame(IWorld world, string title, SaveGame? existingSave, bool autoSave = false, bool quickSave = false)
+    public async Task<SaveGameEvent> WriteSaveGameAsync(IWorld world, string title, IScreenshotGenerator screenshotGenerator, SaveGame? existingSave, bool autoSave = false, bool quickSave = false)
     {
-        if (existingSave == null && autoSave && m_config.Game.RotatingAutoSaves > 0)
-        {
-            var autoSaves = GetSaveGames().Where(x => x.IsAutoSave);
-            var matchingSaves = GetMatchingSaveGames(autoSaves).OrderBy(x => x.Model?.Date);
-            if (matchingSaves.Any() && matchingSaves.Count() >= m_config.Game.RotatingAutoSaves)
-                existingSave = matchingSaves.First();
-        }
-        if (existingSave == null && quickSave && m_config.Game.RotatingQuickSaves > 0)
-        {
-            var quickSaves = GetSaveGames().Where(x => x.IsQuickSave);
-            var matchingSaves = GetMatchingSaveGames(quickSaves).OrderBy(x => x.Model?.Date);
-            if (matchingSaves.Any() && matchingSaves.Count() >= m_config.Game.RotatingQuickSaves)
-                existingSave = matchingSaves.First();
-        }
-        string filename = existingSave?.FileName ?? GetNewSaveName(autoSave, quickSave);
-        var saveEvent = SaveGame.WriteSaveGame(world, title, GetSaveDir(), filename);
+        existingSave = GetExistingSave(existingSave, autoSave, quickSave);
+        bool isNew = existingSave == null;
+        var filename = existingSave?.FileName ?? GetNewSaveName(autoSave, quickSave);
+        var worldModel = world.ToWorldModel();
+        var image = screenshotGenerator.GetImage();
+        var saveEvent = await Task.Run(() => SaveGame.WriteSaveGame(world, worldModel, title, GetSaveDir(), filename, screenshotGenerator, image));
 
+        AddOrUpdateSaveGame(saveEvent.SaveGame);
         GameSaved?.Invoke(this, saveEvent);
         return saveEvent;
     }
 
-    public List<SaveGame> GetSortedSaveGames()
+    public SaveGameEvent WriteNewSaveGame(IWorld world, string title, IScreenshotGenerator screenshotGenerator, bool autoSave = false, bool quickSave = false) =>
+        WriteSaveGame(world, title, screenshotGenerator, null, autoSave, quickSave);
+
+    public SaveGameEvent WriteSaveGame(IWorld world, string title, IScreenshotGenerator screenshotGenerator, SaveGame? existingSave, bool autoSave = false, bool quickSave = false)
+    {
+        existingSave = GetExistingSave(existingSave, autoSave, quickSave);
+        bool isNew = existingSave == null;
+        var filename = existingSave?.FileName ?? GetNewSaveName(autoSave, quickSave);
+        var worldModel = world.ToWorldModel();
+        var saveEvent = SaveGame.WriteSaveGame(world, worldModel, title, GetSaveDir(), filename, screenshotGenerator, screenshotGenerator.GetImage());
+
+        AddOrUpdateSaveGame(saveEvent.SaveGame);
+        GameSaved?.Invoke(this, saveEvent);
+        return saveEvent;
+    }
+
+    private void AddOrUpdateSaveGame(SaveGame newSaveGame)
+    {
+        for (int i = 0; i < m_currentSaves.Count; i++)
+        {
+            var save = m_currentSaves[i];
+            if (save.FileName == newSaveGame.FileName)
+            {
+                m_currentSaves[i] = newSaveGame;
+                return;
+            }
+        }
+
+        m_currentSaves.Add(newSaveGame);
+    }
+
+    private SaveGame? GetExistingSave(SaveGame? existingSave, bool autoSave, bool quickSave)
     {
         var saveGames = GetSaveGames();
-        var matchingGames = GetMatchingSaveGames(saveGames);
-        var nonMatchingGames = saveGames.Except(matchingGames);
-        return matchingGames.Union(nonMatchingGames).ToList();
+        if (existingSave == null && autoSave && m_config.Game.RotatingAutoSaves > 0)
+        {
+            var autoSaves = saveGames.Where(x => x.IsAutoSave).OrderBy(x => x.Model?.Date);
+            if (autoSaves.Any() && autoSaves.Count() >= m_config.Game.RotatingAutoSaves)
+                existingSave = autoSaves.First();
+        }
+
+        if (existingSave == null && quickSave && m_config.Game.RotatingQuickSaves > 0)
+        {
+            var quickSaves = saveGames.Where(x => x.IsQuickSave).OrderBy(x => x.Model?.Date);
+            if (quickSaves.Any() && quickSaves.Count() >= m_config.Game.RotatingQuickSaves)
+                existingSave = quickSaves.First();
+        }
+
+        return existingSave;
     }
 
     public IEnumerable<SaveGame> GetMatchingSaveGames(IEnumerable<SaveGame> saveGames)
@@ -123,10 +156,16 @@ public class SaveGameManager
 
     public List<SaveGame> GetSaveGames()
     {
-        return Directory.GetFiles(GetSaveDir(), "*.hsg")
+        LoadCurrentSaveFiles();
+        m_currentSaves.Sort(new Comparison<SaveGame>(CompareSaveDates));
+        return m_currentSaves;
+    }
+
+    private List<SaveGame> ReadSaveGameFiles()
+    {
+        return [.. Directory.GetFiles(GetSaveDir(), "*.hsg")
             .Select(f => new SaveGame(GetSaveDir(), Path.GetFileName(f)))
-            .OrderByDescending(f => f.Model?.Date)
-            .ToList();
+            .OrderByDescending(f => f.Model?.Date)];
     }
 
     public bool DeleteSaveGame(SaveGame saveGame)
@@ -135,6 +174,8 @@ public class SaveGameManager
         {
             if (File.Exists(saveGame.FilePath))
                 File.Delete(saveGame.FilePath);
+
+            m_currentSaves.Remove(saveGame);
         }
         catch
         {
@@ -146,7 +187,7 @@ public class SaveGameManager
 
     private string GetNewSaveName(bool autoSave, bool quickSave)
     {
-        List<string> files = Directory.GetFiles(GetSaveDir(), "*.hsg")
+        var files = Directory.GetFiles(GetSaveDir(), "*.hsg")
             .Select(Path.GetFileName)
             .WhereNotNull()
             .ToList();
@@ -169,5 +210,16 @@ public class SaveGameManager
         else if (quickSave)
             return $"quicksave{number}.hsg";
         return $"savegame{number}.hsg";
+    }
+
+    private static int CompareSaveDates(SaveGame x, SaveGame y)
+    {
+        if (x.Model == null)
+            return 1;
+
+        if (y.Model == null)
+            return -1;
+
+        return y.Model.Date.CompareTo(x.Model.Date);
     }
 }